backend/requirements.txt
Flask==3.0.3
Flask-PyMongo==3.0.1
python-dotenv==1.0.1
